move : 불필요한 폴더 삭제로 파일 이동
