+++ src/components/mypage/CreatedMeetings.tsx
@@ -2,7 +2,7 @@
 
 import GatheringCard from '@/components/mypage/gatheringCard/GatheringCard';
 import Image from 'next/image';
-import emptyDudu from '../../../../public/images/mypage/dudu-empty.svg';
+import emptyDudu from '../../../public/images/mypage/dudu-empty.svg';
 import Link from 'next/link';
 import { useCreatedGatheringsQuery } from '@/hooks/mypage/queries/useGatheringsQuery';
 import { motion } from 'framer-motion';
