+++ src/components/mypage/Meetings.tsx
@@ -2,7 +2,7 @@
 
 import GatheringCard from '@/components/mypage/gatheringCard/GatheringCard';
 import Image from 'next/image';
-import emptyDudu from '../../../../public/images/mypage/dudu-empty.svg';
+import emptyDudu from '../../../public/images/mypage/dudu-empty.svg';
 import { useJoinedGatheringsQuery } from '@/hooks/mypage/queries/useGatheringsQuery';
 
 export default function Meetings() {
